텍스트 오류 수정

+++ app.py
@@ -166,13 +166,15 @@ def result():
 @app.route('/run_model')
 def run_model():
    mainPy.main()
-   os.rename("static/images/completeImage.png", "static/images/cmImage.png")
+   if(os.path.exists("static/images/completeImage.png")):
+      os.rename("static/images/completeImage.png", "static/images/cmImage.png")
    return render_template('result.html')
 
 @app.route('/result_download')
 def result_download():
    download_dir = "static/images/download"
-   os.rename("static/images/cmImage.png", "static/images/completeImage.png")
+   if(os.path.exists("static/images/cmImage.png")):
+      os.rename("static/images/cmImage.png", "static/images/completeImage.png")
    if not os.path.exists(download_dir):
       os.makedirs(download_dir)
    while(True) :
@@ -186,7 +188,8 @@ def result2():
 
 @app.route('/run_model2')
 def run_model2():
-   os.rename("static/images/completeImage.png", "static/images/cmImage.png")
+   if(os.path.exists("static/images/completeImage.png")):
+      os.rename("static/images/completeImage.png", "static/images/cmImage.png")
    # option_cartoon.html 에서 불러온 option 값
    option = request.args.get("option")
